// src/app/Models/EnfermagemRequest.ts
export interface EnfermagemRequest{
    nome: string;
    dataNascimento: string;
    dataExame: string;
    nivelConciencia: string;
    pupilasSituacao: string;
    pupilaPD: string;
    pupilaPE: string;
    fotorreacao: string;
    sistemaCardioCirculatorio: string;
    carotideosEsquerdo: string;
    carotideosDireito: string;
    braquiaisEsquerdo: string;
    braquiaisDireito: string;
    radiaisEsquerdo: string;
    radiaisDireito: string;
    femoraisEsquerdo: string;
    femoraisDireito: string;
    pediososEsquerdo: string;
    pediososDireito: string;
    popliteosEsquerdo: string;
    popliteosDireito: string;
    perfusaoPeriferica: string;
    edema: string;
    padrao: string;
    murmurioVesicular: string;
    expansaoTorax: string;
    ruidosAdventicios: string;
    aceitacaoDieta: string;
    denticao: string;
    abdome: string;
    rha: string;
    massa: string;
    evacuacao: string;
    genito: string;
    urina: string;
    vesical: string;
    cutanea: string;
    sensorial: string;
    umidade: string;
    atividade: string;
    mobilidade: string;
    nutricao: string;
    friccao: string;
    diagnosticosEnfermagem: string;
    controleSinaisVitaisData: string;
    controleSinaisVitaisHorario: string;
    controleSinaisVitaisPa: string;
    controleSinaisVitaisFc: string;
    controleSinaisVitaisFr: string;
    controleSinaisVitaisTemp: string;
    controleSinaisVitaisPvc: string;
    controleSinaisVitaisGlicemia: string;
    controleSinaisVitaisAnotacoes: string;
}